requirements.txt
asgiref==3.3.1
autopep8==1.5.4
coverage==5.3.1
Django==3.1.4
flake8-django==1.1.1
isort==5.7.0
mccabe==0.6.1
Pillow==8.0.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.4
sqlparse==0.4.1
testfixtures==6.17.1
toml==0.10.2
mysqlclient==2.0.3
django-bootstrap-breadcrumbs==0.9.2
django-mathfilters==1.0.0
django-seed==0.2.2
djangorestframework==3.12.4
channels==2.4.0
channels-redis==2.4.2
